Add basic hash pool
Uses the hashring package, which implements libketama to consistently
hash values on a given host. This is powerful because we can scale out
memcached by adding more hosts and keys are automatically distributed
accross them. At the same time, if a host has problems we can remove
it from the pool and automatically failover to the remaining hosts.